chunk file bug fixes
