+++ src/js/app/Components/index.js
@@ -1,0 +1,9 @@
+import Navbar from "./Navbar";
+import Home from "./Home";
+import Body from "./Body";
+import Profile from "./Profile";
+import Sidebar from "./Sidebar";
+import NavSites from "./NavSites";
+import AFrame from "./AFrame";
+
+export { Navbar, Home, Body, Profile, Sidebar, NavSites, AFrame };
