+++ src/js/app/app.js
@@ -1,15 +1,11 @@
 import React, { useState, useEffect, lazy, Suspense } from "react";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
-import Navbar from "./Components/Navbar";
-import Home from "./Components/Home";
-import Body from "./Components/Body";
-import Profile from "./Components/Profile";
-import Sidebar from "./Components/Sidebar";
-import DataContext from './Utils/DataContext';
-import MenuDataContext from "./Utils/MenuDataContext";
-const NavSites = lazy(() => import("./Components/NavSites"));
-const AFrame = lazy(() => import("./Components/AFrame"));
+import { Body, Home, Navbar, Profile, Sidebar } from "./Components";
+import { DataContext, MenuDataContext } from "./Utils";
+
+const NavSites = lazy(() => import("./Components"));
+const AFrame = lazy(() => import("./Components"));
 
 const appRouter = createBrowserRouter([
   {
@@ -24,7 +20,7 @@ const appRouter = createBrowserRouter([
     children: [
       {
         path: "/",
-        element: <AFrame />,
+        element: <Body />,
       },
       {
         path: "aframe",
